+++ common/src/main/java/com/blib/api/common/event/v1/BLibServerSaveEvent.java
@@ -1,0 +1,9 @@
+package com.blib.api.common.event.v1;
+
+import net.minecraft.server.MinecraftServer;
+
+@FunctionalInterface
+public interface BLibServerSaveEvent {
+
+    void invoke(MinecraftServer server);
+}