+++ common/src/main/java/com/blib/api/common/registry/v1/BLibRegistries.java
@@ -5,9 +5,14 @@ import net.minecraft.resources.ResourceKey;
 import org.jetbrains.annotations.ApiStatus;
 
 import com.blib.api.common.data_sync.v1.model.DataSyncKey;
+import com.blib.api.common.storage.v1.DataStoreType;
 import com.blib.mod.BLib;
 
 public class BLibRegistries {
+
+    public static final ResourceKey<Registry<DataStoreType<?>>> DATA_STORE_TYPES = ResourceKey.createRegistryKey(
+        BLib.MOD.resources().createLocation("data_store_types")
+    );
 
     public static final ResourceKey<Registry<DataSyncKey<?>>> DATA_SYNC_KEYS = ResourceKey.createRegistryKey(
         BLib.MOD.resources().createLocation("data_sync_keys")