+++ common/src/main/java/com/blib/internal/service/BLibEventService.java
@@ -3,11 +3,15 @@ package com.blib.internal.service;
 import org.jetbrains.annotations.ApiStatus;
 
 import com.blib.api.common.event.v1.BLibBlockBreakEvent;
+import com.blib.api.common.event.v1.BLibChunkSaveEvent;
+import com.blib.api.common.event.v1.BLibChunkUnloadEvent;
 import com.blib.api.common.event.v1.BLibCommonSetupEvent;
 import com.blib.api.common.event.v1.BLibEntityTickEvent;
+import com.blib.api.common.event.v1.BLibLevelSaveEvent;
 import com.blib.api.common.event.v1.BLibLevelTickEvent;
 import com.blib.api.common.event.v1.BLibPlayerTrackingEntityEvent;
 import com.blib.api.common.event.v1.BLibServerLifecycleEvent;
+import com.blib.api.common.event.v1.BLibServerSaveEvent;
 import com.blib.api.common.event.v1.BLibTagsUpdatedEvent;
 import com.blib.api.common.event.v1.handle.BLibEventHandle;
 import com.blib.api.common.event.v1.handle.BLibEventListenerHandle;
@@ -16,11 +20,19 @@ import com.blib.api.common.mod.v1.BLibMod;
 @ApiStatus.Internal
 public interface BLibEventService {
 
+    BLibEventListenerHandle<BLibChunkSaveEvent> onChunkSave(BLibMod mod);
+
+    BLibEventListenerHandle<BLibChunkUnloadEvent> onChunkUnload(BLibMod mod);
+
     BLibEventListenerHandle<BLibCommonSetupEvent> onCommonSetup(BLibMod mod);
 
     BLibEventListenerHandle<BLibEntityTickEvent> onEntityTick(BLibMod mod);
 
+    BLibEventListenerHandle<BLibLevelSaveEvent> onLevelSave(BLibMod mod);
+
     BLibEventHandle<BLibPlayerTrackingEntityEvent> onPlayerStartTrackingEntity(BLibMod mod);
+
+    BLibEventListenerHandle<BLibServerSaveEvent> onServerSave(BLibMod mod);
 
     BLibEventHandle<BLibTagsUpdatedEvent> onTagsUpdated(BLibMod mod);
 
@@ -30,11 +42,11 @@ public interface BLibEventService {
 
     BLibEventHandle<BLibLevelTickEvent> preLevelTick(BLibMod mod);
 
-    BLibEventHandle<BLibServerLifecycleEvent.Started> serverStarted(BLibMod mod);
+    BLibEventHandle<BLibServerLifecycleEvent.Started> onServerStarted(BLibMod mod);
 
-    BLibEventHandle<BLibServerLifecycleEvent.Starting> serverStarting(BLibMod mod);
+    BLibEventHandle<BLibServerLifecycleEvent.Starting> onServerStarting(BLibMod mod);
 
-    BLibEventHandle<BLibServerLifecycleEvent.Stopped> serverStopped(BLibMod mod);
+    BLibEventHandle<BLibServerLifecycleEvent.Stopped> onServerStopped(BLibMod mod);
 
-    BLibEventHandle<BLibServerLifecycleEvent.Stopping> serverStopping(BLibMod mod);
+    BLibEventHandle<BLibServerLifecycleEvent.Stopping> onServerStopping(BLibMod mod);
 }